[refactor] extract public contentMetadata generator class
This will make it easier to swap with an implementation that uses cocina
